backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
greenlet>=3.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
alembic>=1.11.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
